@@ -2,9 +2,13 @@ import 'package:flutter/material.dart';
 import 'package:font_awesome_flutter/font_awesome_flutter.dart';
 import 'package:hooks_riverpod/hooks_riverpod.dart';
 import 'package:instagram_clone_course/state/auth/providers/auth_state_provider.dart';
+import 'package:instagram_clone_course/state/image_upload/helpers/image_picker_helper.dart';
+import 'package:instagram_clone_course/state/image_upload/models/file_type.dart';
+import 'package:instagram_clone_course/state/posts_settings/providers/post_settings_provider.dart';
 import 'package:instagram_clone_course/views/components/dialogs/alert_dialog_model.dart';
 import 'package:instagram_clone_course/views/components/dialogs/logout_dialog.dart';
 import 'package:instagram_clone_course/views/constants/strings.dart';
+import 'package:instagram_clone_course/views/create_new_post/create_new_post_view.dart';
 import 'package:instagram_clone_course/views/tabs/users_posts/user_posts_view.dart';
 
 // when you are already logged in
@@ -30,13 +34,55 @@ class _MainViewState extends ConsumerState<MainView> {
               icon: const FaIcon(
                 FontAwesomeIcons.film,
               ),
-              onPressed: () async {},
+              onPressed: () async {
+                // pick a video first
+                final videoFile =
+                    await ImagePickerHelper.pickVideoFromGallery();
+                if (videoFile == null) {
+                  return;
+                }
+                ref.refresh(postSettingProvider);
+                // go to the screen to create a new post
+                if (!mounted) {
+                  return;
+                }
+                Navigator.push(
+                  context,
+                  MaterialPageRoute(
+                    builder: (_) => CreateNewPostView(
+                      fileToPost: videoFile,
+                      fileType: FileType.video,
+                    ),
+                  ),
+                );
+              },
             ),
             IconButton(
               icon: const Icon(
                 Icons.add_photo_alternate_outlined,
               ),
-              onPressed: () async {},
+              onPressed: () async {
+                // pick an image
+                final imageFile =
+                    await ImagePickerHelper.pickImageFromGallery();
+                if (imageFile == null) {
+                  return;
+                }
+                ref.refresh(postSettingProvider);
+                // go to the screen to create a new post
+                if (!mounted) {
+                  return;
+                }
+                Navigator.push(
+                  context,
+                  MaterialPageRoute(
+                    builder: (_) => CreateNewPostView(
+                      fileToPost: imageFile,
+                      fileType: FileType.image,
+                    ),
+                  ),
+                );
+              },
             ),
             IconButton(
               icon: const Icon(
